Spectra/Magellan hardware split into multiple files.
