+++ paradox/hardware/evo/__init__.py
@@ -1,1 +1,3 @@
 from .panel import Panel
+
+__all__ = ['Panel']
